Modulo15 - separando lógica de controller em services
